新增配件商品

+++ src/views/pages/admin/components/goods.vue
@@ -1,14 +1,25 @@
 <template>
-  <div class="goods" v-loading='loading'>
+  <div class="goods" v-loading="loading">
     <div class="title">已上架商品</div>
-    <el-table :data="tableData" border stripe @selection-change="handleSelectionChange"  :row-class-name="tableRowClassName">
+    <el-table
+      :data="tableData"
+      border
+      stripe
+      @selection-change="handleSelectionChange"
+      :row-class-name="tableRowClassName"
+    >
       <el-table-column type="selection" width="55"></el-table-column>
       <el-table-column label="商品名称" prop="name"></el-table-column>
       <el-table-column label="商品品牌" prop="brand"></el-table-column>
       <el-table-column align="right">
-        <template slot="header" >
-          <el-input v-model="key" size="mini" placeholder="输入关键字搜索" clearable    @keyup.native="getData"
-            debounce/>
+        <template slot="header">
+          <el-input
+            v-model="key"
+            size="mini"
+            placeholder="输入关键字搜索"
+            @keyup.native="getData"
+            debounce
+          />
         </template>
         <template slot-scope="scope">
           <div class="row-ctrl">
@@ -21,7 +32,11 @@
     </el-table>
     <div class="btns">
       <el-button type="text" @click="addGoods">上架新商品</el-button>
-      <el-button type="text" @click="deleteSelestions" :disabled="!selections.length">下架所选商品</el-button>
+      <el-button
+        type="text"
+        @click="deleteSelestions"
+        :disabled="!(selections.part.length||selections.good.length)"
+      >下架所选商品</el-button>
       <el-button type="text" @click="deleteAll" :disabled="!tableData.length">下架所有商品</el-button>
     </div>
     <el-pagination
@@ -51,8 +66,11 @@ export default {
       total: 40,
       tableData: [],
       key: "",
-      selections: [],
-      loading:false
+      selections: {
+        good: [],
+        part: []
+      },
+      loading: false
     };
   },
   components: {
@@ -63,18 +81,15 @@ export default {
     this.getData();
   },
   methods: {
-    tableRowClassName({row})
-    {
-      if(row.state=='warning')
-      {
-return 'warning-row'
-      }
-      else{
-        return ''
+    tableRowClassName({ row }) {
+      if (row.state == "warning") {
+        return "warning-row";
+      } else {
+        return "";
       }
     },
     getData() {
-      this.loading=true
+      this.loading = true;
       this.$axios({
         url: "/ego/good/all",
         params: {
@@ -83,25 +98,27 @@ return 'warning-row'
           key: this.key
         }
       }).then(res => {
-        this.loading=false
+        this.loading = false;
         this.tableData = this.handleData(res.data.data);
         this.total = res.data.total;
       });
     },
-    handleData(data)
-    {
-      for(var n=0;n<data.length;n++)
-      {
-        var config = data[n].config
-        for(var j=0;j<config.length;j++)
-        {
-          if(config[j].number<10)
-          {
-            data[n].state = 'warning'
+    handleData(data) {
+      for (var n = 0; n < data.length; n++) {
+        var config = data[n].config;
+        if (config) {
+          for (var j = 0; j < config.length; j++) {
+            if (config[j].number < 10) {
+              data[n].state = "warning";
+            }
+          }
+        } else {
+          if (data[n].number < 10) {
+            data[n].state = "warning";
           }
         }
       }
-      return data
+      return data;
     },
     handleExplain(data) {
       this.$refs.explainDialog.controlDialog({
@@ -110,36 +127,45 @@ return 'warning-row'
       });
     },
     deleteSelestions() {
-              this.$confirm('确定下架所选商品吗?', '提示', {
-          confirmButtonText: '确定',
-          cancelButtonText: '取消',
-          type: 'warning'
-        }).then(() => {
-          this.loading=true
-      this.$axios({
-        url: "/ego/good/deleteByIds",
-        method: "post",
-        data: {
-          ids: this.selections
-        }
-      }).then(res => {
-        if (res.data.code) {
-          this.loading=false
-          this.getData();
-          this.$message({
-            type: "success",
-            message: "下架成功!"
+      this.$confirm("确定下架所选商品吗?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      })
+        .then(() => {
+          this.loading = true;
+          this.$axios({
+            url: "/ego/good/deleteByIds",
+            method: "post",
+            data: {
+              ids: this.selections
+            }
+          }).then(res => {
+            if (res.data.code == 1) {
+              this.loading = false;
+              this.getData();
+              this.$message({
+                type: "success",
+                message: "下架成功!"
+              });
+            }
+            if (res.data.code == 3) {
+              this.loading = false;
+              this.getData();
+              this.$message({
+                type: "warning",
+                message: res.data.msg
+              });
+            }
           });
-        }
-      });
-        }).catch(() => {
-          this.loading=false
+        })
+        .catch(() => {
+          this.loading = false;
           this.$message({
-            type: 'info',
-            message: '已取消下架'
-          });          
+            type: "info",
+            message: "已取消下架"
+          });
         });
-
     },
     deleteAll() {
       this.$confirm("确认下架所有商品吗?", "提示", {
@@ -148,23 +174,32 @@ return 'warning-row'
         type: "warning"
       })
         .then(() => {
-          this.loading=true
+          this.loading = true;
           this.$axios({
             url: "/ego/good/deleteAll"
           }).then(res => {
             if (res.data.code == 1) {
               this.currentPage = 1;
-              this.loading=false
+              this.loading = false;
               this.getData();
               this.$message({
                 type: "success",
-                message: "下架成功!"
+                message: res.data.msg
+              });
+            }
+            if (res.data.code == 3) {
+              this.currentPage = 1;
+              this.loading = false;
+              this.getData();
+              this.$message({
+                type: "warning",
+                message: res.data.msg
               });
             }
           });
         })
         .catch(() => {
-          this.loading=false
+          this.loading = false;
           this.$message({
             type: "info",
             message: "已取消下架"
@@ -182,7 +217,7 @@ return 'warning-row'
     handleEdit(data) {
       this.$refs.modifyDialog.controlDialog({
         flag: true,
-        data:{...data},
+        data: { ...data }
       });
     },
     handleDelete(data) {
@@ -192,26 +227,31 @@ return 'warning-row'
         type: "warning"
       })
         .then(() => {
-          this.loading=true
+          this.loading = true;
           this.$axios({
             url: "/ego/good/deleteById",
             method: "post",
-            data: {
-              _id: data._id
-            }
+            data
           }).then(res => {
-            if (res.data.code) {
-              this.loading=false
+            if (res.data.code == 1) {
+              this.loading = false;
               this.getData();
               this.$message({
                 type: "success",
                 message: "下架成功!"
               });
+            } else if (res.data.code == 3) {
+              this.loading = false;
+              this.getData();
+              this.$message({
+                type: "success",
+                message: res.data.msg
+              });
             }
           });
         })
         .catch(() => {
-          this.loading=false
+          this.loading = false;
           this.$message({
             type: "info",
             message: "已取消下架"
@@ -219,9 +259,16 @@ return 'warning-row'
         });
     },
     handleSelectionChange(selections) {
-      var arr = [];
+      var arr = {
+        good: [],
+        part: []
+      };
       for (var n = 0; n < selections.length; n++) {
-        arr.push(selections[n]._id);
+        if (selections[n].config) {
+          arr.good.push(selections[n]._id);
+        } else {
+          arr.part.push(selections[n]._id);
+        }
       }
       this.selections = arr;
     },
